BackEnd/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose
requests
python-multipart
pdfminer.six
docx2txt
redis
email-validator
google-generativeai
python-dotenv
PyMuPDF
python-docx
asyncio
# elasticsearch==8.11.0  # DISABLED - Elasticsearch not in use